refactor: profile page
- 각 section 별로 content 컴포넌트로 분리
- 컴포넌트 분리에 따른 구조 수정
- 몇 가지 주석 추가
- 로딩 처리를 위해 suspense 코드 추가
- notification 에러 디버깅: useSession() => auth()로 수정
