+++ src/shared/hooks/use-user.ts
@@ -1,7 +1,7 @@
-import { useSession } from 'next-auth/react'
+import { auth } from '@/app/api/auth/[...nextauth]/auth'
 
-export const useUser = () => {
-  const { data: session } = useSession()
+export const useUser = async () => {
+  const session = await auth()
   const user = session?.user.dto
 
   return { user }
